pyproject.toml
[project]
name = "nnrika"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.4.0",
    "einops>=0.8.1",
    "google-genai>=1.5.0",
    "httpx[socks]>=0.28.1",
    "pillow>=11.0.0",
    "pydantic-settings>=2.8.1",
    "pyvips>=2.2.3",
    "requests[socks]>=2.28.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.48.3",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/nightly/cu128"
default = false

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "python-lsp-server[all]>=1.12.2",
    "ruff>=0.9.10",
]
